requirements.txt
APScheduler==3.11.2
asyncpg==0.30.0
aiosqlite==0.21.0
alembic==1.16.4
anyio==4.13.0
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
numpy==2.4.2
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.20
pytz==2026.1.post1
requests==2.32.5
sentry-sdk==2.35.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
structlog==25.4.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.35.0
watchdog==6.0.0
